Add files via upload
Note: this script was made on a mac so when using this on windows there will be a few things you will need to change. For instance, make sure that you change the file path so that it can be accessed by a windows machine. If you need help you can reach out to me.